Use kankyo to load environment variables
kankyo is a crate that unloads environment variables listed in the '.env' file into `std::env`, which can be accessed via `std::env::var` like other environmental variables.

This allows the user to separate sensitive login data from other data that is, perhaps, more generalised. However, this adds a dependency for something simple while continuing to use a 'config.toml' setup...is this worth it?

Lastly, additions were added to the README to specify the locations of the config and env files, as well as adding '.env' to the .gitignore.